+++ pennylane_lightning/src/rework/Gates.cpp
@@ -422,3 +422,44 @@ unique_ptr<Pennylane::AbstractGate> Pennylane::constructGate(const string& label
 
     return dispatchTableIterator->second(parameters);
 }
+
+void Pennylane::constructAndApplyOperation(
+    StateVector& state,
+    const string& opLabel,
+    const vector<unsigned int>& opWires,
+    const vector<double>& opParams,
+    const unsigned int qubits
+) {
+    unique_ptr<AbstractGate> gate = constructGate(opLabel, opParams);
+    if (gate->numQubits != opWires.size())
+        throw std::invalid_argument(string("The gate of type ") + opLabel + " requires " + std::to_string(gate->numQubits) + " wires, but " + std::to_string(opWires.size()) + " were supplied");
+
+    const vector<CplxType>& matrix = gate->asMatrix();
+
+    vector<size_t> internalIndices = generateBitPatterns(opWires, qubits);
+
+    vector<unsigned int> externalWires = Pennylane::getIndicesAfterExclusion(opWires, qubits);
+    vector<size_t> externalIndices = generateBitPatterns(externalWires, qubits);
+
+    vector<CplxType> inputVector(internalIndices.size());
+    for (const size_t& externalIndex : externalIndices) {
+        CplxType* shiftedStatePtr = state.arr + externalIndex;
+
+        // Gather
+        size_t pos = 0;
+        for (const size_t& internalIndex : internalIndices) {
+            inputVector[pos] = shiftedStatePtr[internalIndex];
+            pos++;
+        }
+
+        // Apply + scatter
+        for (size_t i = 0; i < internalIndices.size(); i++) {
+            size_t internalIndex = internalIndices[i];
+            shiftedStatePtr[internalIndex] = 0;
+            size_t baseIndex = i * internalIndices.size();
+            for (size_t j = 0; j < internalIndices.size(); j++) {
+                shiftedStatePtr[internalIndex] += matrix[baseIndex + j] * inputVector[j];
+            }
+        }
+    }
+}
